backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.11
autopep8==2.3.2
click==8.2.1
dill==0.4.0
exceptiongroup==1.3.0
fastapi==0.116.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
packaging==25.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pylint==3.3.7
sniffio==1.3.1
starlette==0.47.2
tomli==2.2.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
passlib[bcrypt]
python-jose[cryptography]